pyproject.toml
[project]
name = "mcp-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "isort>=6.0.1",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
]
